Rename package to html_table. Change api to take *html.Node.
See, coming up with a decent name is so much easier to do after writing
the code in question. After those 4 minutes, the best runner-ups I had
were "iterator" and "iterate". And I wasted 4 minutes for nothing.
Make the client parse the html first; it's easier for it to handle
errors in place, and it avoids parsing the same html multiple times.