Work on sorting imports in InlineDotImports().
It's really bad. Editing an existing AST is near impossible, because I
need to be able to insert a newline, but that requires modifying the
token.FileSet, which only has public API for merging lines, not adding
new ones. I end up having to print AST, modify the text, and
(optionally) parse to get expected AST back. But it works and passes a
test for now.