Remove unused experimental "path" package.
This was an experiment to represent paths in a structured manner rather
than as a string.
I didn't like the `Relative bool` aspect of this approach, since it
doesn't scale to paths other than relative to cwd or relative to root.
My new insight is that all paths are always relative, it's a matter of
what they're relative to. That may lead to a better way to do this.
Remove this package since no one imports it and I can rewrite it if I
decide to revisit this.