Workaround OS X ACPI crash on suspend/resume
This may fix/avoid the OSX ACPI/Hypervisor framework crash issue without
rebooting the VM. I've run it overnight on two laptops with 10.11.2 and 10.11.3
without any crashes and tested it 10+ times with short hibernation timeouts and
sleep/wakeup. However, since we don't have a test that can reproduce this issue
all the time it needs testing on more systems to be sure that it actually
works.

Freeze is implemented by adding a condition variable in the hv_vcpu_run
interrupt loop, which each vcpu thread will wait on if SIGUSR1 has been
received. Before waiting the vcpu is set to a frozen state and memory is
protected by setting it -rwx. The memory protection may not be necessary, but
I'm hoping it will give us a better error message if the crash happens again.
The theads are resumed on SIGUSR2.

This currently requires an external agent to send the SIGUSR{1,2} upon
{suspend,resume}. Perhaps an event listening for power events should be used
instead.

Signed-off-by: Magnus Skjegstad <magnus@skjegstad.com>