prevent spinning if the tty is disconnected
When a slave disconnects, we get an infinite stream of read events
on the master side of the pty, where a `read` will return 0 i.e.  end-of-file.
This patch suppresses these events by always keeping a slave fd open: when
end-of-file is received, we close and re-open our slave fd. We never read from
the slave fd, but its mere presence is enough to stop hyperkit spinning.

Signed-off-by: David Scott <dave.scott@docker.com>