Rename u/u11 to ioutil.
This is an improvement over previous name, but unsure if shadowing
io/ioutil package name is a good idea. Give it a try and see how it
goes.

Change parameter order in WriteFile to be more consistent with
"io/ioutil".WriteFile. Not sure why it wasn't that way to begin with.