Delete last gist package.
Its code is now used directly by the one command that needs this
functionality. That allows to simplify it and specialize for that
command's needs. It's not helpful to have this as a general package for
anyone to use.

This is the last Go package with "gist" in the name! :fire: Woohoo,
finally no more gist packages! :tada: