Delete html_table, html_to_markdown packages.
These packages were created a long time ago, for a specific task. It
was to parse some information out of HTML documents.

In an ideal situation, there should be a better source information in a
more standard, machine-readable format like JSON, etc., and it
shouldn't be necessary to be parsing HTML tables.

Delete these packages since they're unused (no known public importers).
If this functionality is ever needed in the future, it can be brought
back via git history.

These packages are not very idiomatic and would need work to be brought
up to standards. However, without a direct use case, it's hard, so
better to simply delete them.