vsock: use lists for free/inuse sockets
This reduces array traversals where we need to lock each element in turn.

Since both the rx and tx threads need to traverse the list they both need to
take the list lock, which introduces more contention, make it an rw lock which
in the common case is taken for read only in both threads and only taken for
writing in the socket open and close cases. The close case needs to be
deferred via a queue in the rx thread to avoid needing to upgrade the read lock
to a write lock.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>