Introduce simple boot cycle test for hyperkit.
Fixes #24

Adds dependency on expect, wired in through Makefile.

The expect script replicates the command from hyperkitrun.sh to prevent
requiring a prompt whilst in expect. The dependency chain is handled in
the Makefile.

Tested both success and failure using make test - adjusting the timeout
to 1 will give an example fail that propgates up to the make return.

Signed-off-by: Pris Nasrat <pnasrat@gmail.com>