vsock: check fd's are selectable as necessary when openining
Rather than checking fd's vs FD_SETSIZE as we construct the fdsets instead do
it when we open them. For global fds (kick pipes, connect socket, forwarding
fd) which are created at start of day this is (still) a hard failure (but
sooner, and with more specific logging). For fds associated with socket
connections we refuse the connection attempt, like any other failure on that
path.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>