firmware/multiboot: implement full support for multiboot protocol
The patch implements full support of multiboot protocol (kernel +
modules [with arguments] + command line). With this patch it's possible
to boot, for example, SmartOS using the following -f argument:

"multiboot,\
/Volumes/CDROM/platform/i86pc/kernel/amd64/unix,\
/Volumes/CDROM/platform/i86pc/amd64/boot_archive:\
/Volumes/CDROM/platform/i86pc/amd64/boot_archive.hash,\
-B smartos=true,console=ttya"

It starts to boot but later hyperkit crashes with vm exit/exit_reason=33
(INVAL_VMCS) while doing int 0x1a/ax=b101. Obviously, this is because
there is not BIOS and has nothing to do with multiboot itself.

Signed-off-by: Pavel Borzenkov <pavel.borzenkov@gmail.com>