gopherjs_http: Skip .go files that don't belong.
Make note that current behavior, while better than before, is still
wrong, because build tags are disregarded. This can cause unnecessary
things to be built and included in the output.

This GopherJS compile-in-memory VFS functionality will eventually be
merging with the GopherJS one, and this issue will be resolved at that
time.