Add a `platform_locks.h` which lets us switch between old/new locking
The old locking scheme using OSSpinLockLock is deprecated as of
OSX 10.12, but the new one is only available in OSX 10.12 onwards,
so we need to support both to transition.

With this patch, the code will compile cleanly on 10.12, but will
fail if MACOSX_DEPLOYMENT_TARGET is set to an earlier version (due
to Werror on the deprecation warning). This can be addressed in a
future patchset, as this lets us test the new spinlocks on OSX 10.12
beta right now.

Signed-off-by: David Wisner <dwisner6@gmail.com>