vsock: Add some comments and descriptively named macros to reply ring
This (hopefully) helps make it clearer what is going on when the ring has
filled up and why the various tests are correct.

This doesn't actually fix docker/hyperkit#50 but will make an eventual fix less
brain bending to think about.

Note that REPLY_RING_FULL is not actually used and is just for documentation
purposes. Comment it out to keep -Wunused-macros happy.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>