Fix local shutdown
The `assert` and the` _rx` suffix on the function are wrong, it is perfectly
fine to call from either rx or tx thread.

In principle it would be possible to have an optimised tx-thread-only version
which could skip the CLOSING_TX state and go right to CLOSING_RX, the current
state prior to this commit is a botched/incomplete attempt at such an
optimisation (but the wrong way round and broken).

Signed-off-by: Ian Campbell <ian.campbell@docker.com>