gopherjs_http: Minor cleanup, print minify option value.
The goal was to make it not minify when in development mode (i.e.,
'dev' build tag is used), since it's may be helpful to see verbose JS
output.

However, I couldn't get that to work for now because go generate ./...
may use vfsgendev which specifically uses 'dev' build tag when
generating.

Need to find a way to resolve that problem before I can proceed. For
now, just include the minor style improvements.