vsock(v7): Ensure we periodically check the reply ring even with lots of rx
v7 of the virtio vsock spec proposal adds a requirement to continue processing
the TX ring so long as "additional resources" are available in order to avoid
deadlock. In this implementation those additional resources take to form of the
reply_ring.

In order to avoid deadlock we must ensure that we process the reply ring in
preference to actual data RX frequently enough that we avoid filling the reply
ring too. Since the reply ring is 2x the virtio ring size I _think_ it should
be sufficient to ensure that we check for new reply entries periodically
(something less than the virtio ring size) rather than every time.

This is discussed in several replies to the spec posting at
<1470324277-19300-1-git-send-email-stefanha@redhat.com>.

A more formal consideration of how deadlock is avoided using this scheme should
be forthcoming (see the above thread). Worst case is this change doesn't avoid
deadlock, it can't make things any worse.

This should hopefully address docker/hyperkit#50.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>