Add block_delete to handle BOP_DELETE / TRIM
Previously we would return EOPNOTSUPP on non-FreeBSD in response to
a BOP_DELETE. This patch calls a new `block_delete` function instead
which calls `DIOCGDELETE` on FreeBSD as before and will call some
Mirage block API in future.

Note that no behaviour has changed because the device still does not
advertise the capability (`candelete` is false) so this function will
still return `EOPNOTSUPP`.

Signed-off-by: David Scott <dave.scott@docker.com>