vsock: If we fail to open a socket then put it back on the free list
If we alloc_sock and then error out without actually connecting (e.g. we fail
to open the Unix socket) we need to make sure we put the socket struct back
onto the free list, rather than marking it FREE but leaving it on the active
list.

Refactor the relevant parts of pci_vtsock_rx_thread() (which deals with freeing
sockets which were successfully opened) into a new function and call it on the
error path wherever we call alloc_sock().

Signed-off-by: Ian Campbell <ian.campbell@docker.com>