ocaml: if a Mirage thread fails, ensure we fail the request
The Mirage I/O operations are supposed to return a result type as follows:

  [ `Ok of result | `Error of error description ]

If the implementation actually fails i.e. by raising an exception, this would
propagate and prevent the client request ever returning. This patch catches
stray exceptions, logs something about them, and returns an error to the
caller.

Signed-off-by: David Scott <dave.scott@docker.com>