vsock: add an AF_UNIX passed-fd-based vsock configuration interface
If messages with attached fds are sent to an AF_UNIX socket that proxies
an AF_VSOCK connection, they are treated as configuration RPC
requests. RX thread activity will block until the configuration request
is applied. The passed fd is valid for a single use in which the
requesting process writes a configuration message and, if appropriate, a
reply is written back. Once the configuration request is processed, the
fd is closed.

This commit introduces version 1 of the configuration protocol that
currently has a message format that looks like:

struct vtsock_config_hdr {
       uint32_t len;
       uint32_t version;
       uint32_t type;
};

Signed-off-by: David Sheets <dsheets@docker.com>