circle.yml: detect if the CI base image has the qcow libraries installed
The CI first builds and tests without qcow, and then with qcow. If the
CI base image were to be modified to have the qcow dependencies installed
then the CI would build with qcow twice and we might not notice.

This patch checks that a qcow-related build artifact is not created after
the without-qcow build.

Signed-off-by: David Scott <dave.scott@docker.com>