make: separate the qcow test into a `test-qcow` target
Previously `make test` would run the without-qcow test and then the
with-qcow test if it detected the libraries. Unfortunately the libraries
will only be detected if we've set the opam environment variables
properly.

This patch makes the qcow testing more explicit by calling for it
directly in the circle.yml.

Signed-off-by: David Scott <dave.scott@docker.com>