go: Make sure we reap child hyperkit processes if backgrounded
In some use cases, the go bindings may be used to start
hyperkit processes in the background and the processes may
get killed externally. reaping the child process makes
sure that there are no zombie processes.

An alternative would be to double fork(), but that seems a bit
iffy in Go.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>