go lib: Set vsock guest CID and default ports to forward
Adds support for setting the remaining virtio-sock/vsock parameters from
the go library.

`VSockPorts` specifies the vsock guest ports that will be forwarded
automatically on startup back to the host. A unix domain socket will
be created in the state directory for each port. Also exposed in
the cli sample as `-vsock-ports`

`VSockGuestCID` sets the guest CID. It defaults to 3 by default (as
before).

Signed-off-by: Magnus Skjegstad <magnus@skjegstad.com>