go lib: Add support for 9p sockets
Add support for specifying the path and tag of 9p unix domain sockets
that will be exposed to the guest. This is needed to e.g. expose the
vpnkit /ports directory in the guest.

Also adds a nextSlot counter when building the command line so it's
easier to add new parameters that could require multiple PCI slots.

The sample only supports a single 9p socket for now.

Signed-off-by: Magnus Skjegstad <magnus@skjegstad.com>