go: Rework the Console handling code
The hyperkit code expects a pty when comX is connected to
stdio. This does not easily allow for redirecting the output
to a file or pipe, which is useful for testing.

This commit uses the PTY mode for comX if the stdout is not
a terminal and the StateDir is defined (to hold the PTY slave).
It then uses a go routine to copy the output from the VM to
stdout allowing it to be redirected. stdin is ignored for simplicity.
If one also wants to redirect stdin, it's best to use 'expect' or
'script'

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>