go: Make hyperkit wrapper compilable on other OSes
In LinuxKit we want to use the HyperKit wrapper on all supported
OSes and just error out when not running on macOS.

The previous commit (terminal handling) introduced macOS specifics
which fail to compile for Linux. This commit adds a fallback
for OSes other than macOS, which simply error out when called.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>