go: Add cross-compile tests for Linux and Windows
While the hyperkit go package is only useful on macOS, consumers
may want to use it in code which also also needs compiling on Windows
and Linux without having to use OS conditional compilation.

This patch adds checks to the CI to ensure the code cross compiles
on other platforms.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>