gopherjs_http/jsutil: Document panic behavior when argument count mismatch.
It's expected that you call the wrapped func with the correct number of
arguments (i.e., same as the original func expects). If you don't, the
code previously panicked in a way that was hard to tell what the
problem was.

Document this behavior and make the panic friendlier for easier
debugging.