generated: Remove comment about API being undecided.
Enough time has passed since the API was created, and so far, I have
not had desires to change it, nor has anyone else suggested any API
improvements. It's likely going to stay as is, so remove the comment.
If there are good change ideas in the future, it can still be
considered. But the comment was meant for the very early days of the
package's existence, and has outlived its usefulness.

Also make the comment more logical by using double quotes for a string
with 2 characters.