Update to vpnkit protocol 22 and add preferred_ipv4 option
This updates the vpnkit driver to the latest protocol version. The new
protocol allows the client to request an IPv4 address without encoding
it in the UUID and supports returning error messages from the server.

A new parameter to the VPNKit driver, "preferred_ipv4", allows a
specific IP address within the range to be requested. The IP address is
assigned to the VPNKit UUID. The same UUID has to be used to get the
same IP after it has been assigned.

This change is incompatible with older versions and will only work with
recent versions of VPNKit.

Signed-off-by: Magnus Skjegstad <magnus@skjegstad.com>