go: don't check url-like paths of disks
This is arguably not the proper way to address this issue: we need to
extend the support for more disk format so that the caller can pass
the real path, and we would assemble the url here.

This will be done in the near future.  Meanwhile this commit allows
pinata to use hyperkit.go.

Signed-off-by: Akim Demaille <akim.demaille@docker.com>