go: define methods on disks
Other disk image format will be provided (Qcow2).  In other to
manipulate disks independently of their format, we will introduce an
interface and two structs that implement this interface.

Interfaces in Go are not allowed to have attributes (such as Path), so
we need accessors (such as SetPath and GetPath).

This commit changes the DiskConfig struct with free functions orbiting
around it into one with methods.

Signed-off-by: Akim Demaille <akim.demaille@docker.com>