go: scope the logger to the whole module
Currently the logger is bound to an HyperKit struct.  But there are
places where we would like to log which are not bound to any such
struct.

So lift the logger to the module level.

Also, remove support for Panicf (we don't plan to quit on the caller
like this), and introduce Debug and Info levels.

Signed-off-by: Akim Demaille <akim.demaille@docker.com>

fixup! go: scope the logger to the whole module