go: allow to capture errors in background executions
Currently we offer two means to run hyperkit: Start and Run.  Both run
the process in background (using exec.Cmd.Start), in order to be able
to save the state when launching succeeded.

Both wait for the command to finish.  Run returns the exit status.
Start waits in a goroutine and discard the result.

As a result, the user has no clean means to run in background and get
the exit status.  The `Start` function discards the exit status.  But
since `Run` is blocking, the caller must run it in a goroutine.  But
then the callers needs to watch the process to know when it is
actually running (e.g., by watching h.Pid).

Rewrite this:
- the HyperKit struct now features a channel to publish the
  exit status of hyperkit.
- let Start implement the full background scenario, subsuming
  what execute and execHyperKit used to do.
- let Run be implemented on top of Run.
- HyperKit.background is now useless.

Signed-off-by: Akim Demaille <akim.demaille@docker.com>