Go: remove the ExtraFiles field
It was introduced for Docker for Mac in
d1a5c32a404a86facc5800d962901b5e7a1783d9, so that we could ensure that
Docker for Mac's lockfile would be preserved even if it were to crash.
This use was removed in 2bedc1969f8c6893678fdef55e90e734c170c42c since
djs55 had qcow-tool deal with the lock itself.

Signed-off-by: Akim Demaille <akim.demaille@docker.com>