Update circle config to 2.0
Somewhat based on https://circleci.com/docs/2.0/migrating-from-1-2/, but needed
some other manual tweaks not mentioned there.

This gives us access to some of the new/more powerful features, but the only
things we use it for here is to move the artifacts to the top level avoiding
Needing the click/expand through `Container 0` → `Users/` → `distiller/` →
`hyperkit/` → `build/`.

opam appears to error out (having successfuly processed depexts) if `$TERM` is
not set (and circleci doesn't seem to set it) so manually set it to `vt100`.

Signed-off-by: Ian Campbell <ijc@docker.com>