Fix TRIM/DISCARD of small regions
We present a virtual sector size of 512 to the VM but are likely using
a 4096 byte native sector size on the local SSD, therefore we have to
take care to align calls to `fcntl(F_PUNCHHOLE)`.

- Cap the size of the leading sectors which we zero to `fp_length`
  (otherwise we can zero sectors outside the range to be trimmed)
- Don't call `fcntl(F_PUNCHHOLE)` with a `fp_length` of zero as this will
  return `EINVAL` (and is pointless)
- Re-order the calls to `pwrite` and `fcntl(F_PUNCHHOLE)`:
  previously we would `pwrite` the leading and trailing sectors and then
  punch out the middle. Now we walk through the region, updating `fp_offset`
  and `fp_length` as we go. This is hopefully clearer.
- Add more linebreaks to make the code more readable

For example this fixes the following case, when `fp_length` is small:

assume `fp_offset = 1024` and `fp_length = 512` (i.e. TRIM one sector)
=> `aligned_offset = 4096`
=> `len_to_zero = aligned_offset - fp_offset = 3072`

and we will erase `3072` bytes even though the request was for `512`.
We will then go on to calculate a -ve number for the `fcntl(F_PUNCHHOLE)`
call which will return `EINVAL`, but damage has already been done.

Thanks to @rn for his assistance with this!

Signed-off-by: David Scott <dave.scott@docker.com>