Offer a means to send the tty to the logs
On Docker for Windows, all the logs are aggregated: GUI, LinuxKit,
etc.  Currently debugging on Docker for Mac is less easy, in
particular because LinuxKit's log are sent to the consolefile rings of
fixed sized buffers.  As a result, it's hard to find the interleaving
of actions between Docker for Mac and LinuxKit, the console files are
fragmented and full of '\0', etc.

It is possible to have hyperkit redirect the tty's output to its
stderr/stdout, and then its caller (Docker for Mac's driver) could
redirect these streams to the logger.  Unfortunately on recent macOS
releases, asl is replaced by os_log which uses exclusively the path of
the executable as source (changing `argv[0]` is useless).  As a
result, hyperkit's logs are "credited" to the driver.

The most elegant approach is to embrace asl/os_log.

Currently Docker for Mac still targets 10.11, and os_log appeared in
macOS 10.12, so we cannot use directly os_log's API, we use it through
the asl shim.

Add a new device configuration, asl, that sends the tty's output to asl.

Signed-off-by: Akim Demaille <akim.demaille@docker.com>