tests: wait for hyperkit to exit before examining the disk image
We were seeing cases in CI where the subsequent examination of the disk was
failing with `hdiutil: mount failed - no mountable file systems`.

It was hypothesised (and tested in #212 and #214) that perhaps Hyperkit was not
completely shutting down (and flushing any outstanding I/O) before the script
was examining the disk.

Add an `expect` call to both tests which waits for `eof` from the fds attached
to the Hyperkit process (`eof` here means the other end has shutdown). In the
qcow case this replaces the existing `close` since `expect` does a close in the
`eof` case already.

Once Hyperkit has exited we can `wait` to reap the process, this might be
unnecessary since the test script isn't far from exiting itself which will end
up collecting all orphaned children but we might as well and it gives some
extra confidence that Hyperkit really has gone away.

To help diagnose any future problems we also collect both disk images as
artifacts (they are reasonably small). To ensure that `disk.dmg` is always
stashed, not just on failure, remove any existing `disk.dmg` at the start
(`hdiutil create` insists that the file does not exist) rather than the end.

Signed-off-by: Ian Campbell <ijc@docker.com>