Open ISO images (ahci-cd backing files) as read-only
Commit 6e72b75 ("Lock the backing file to prevent concurrent
modification") added locking to block device backing files. It
uses a shared lock for read-only devices (',ro' option) and an
exclusive lock by default.

The ahci-cd device only supports read-only ISOs as backing file
so it is cumbersome having to specify ',ro' as option in order
to share the same ISO between different HyperKit instances.

This commit adds a 'ro' argument to 'blockif_open()' and only
sets it for ahci-cd devices. This then causes the ISO backing
file to opened with 'O_RDONLY' and a shared lock being taken.
There should be no change for other block devices.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>