ocaml: be compatible with -safe-string
In newer versions of OCaml (4.06 and later) -safe-string is on by
default which means there is a difference between the `string` and
`bytes` types where `string` is immutable but `bytes` is mutable.

The expression `Bytes.unsafe_of_string "X"` returns a mutable reference
to the constant buffer "X" but the `Unix.write` call will not write to
it.

Signed-off-by: David Scott <dave.scott@docker.com>