Implement PCI pre-initialisation with privilege dropping
This implements the scaffolding for PCI devices to perform early
initialisation which requires root privileges, so that hyperkit can
be installed setuid. It will then drop privileges back to the
calling user after performing pre-initialisation.

This also adds an fi_arg member to struct funcinfo, which allows a
pe_preinit implementation to pass a pointer as an argument to
pe_init.

An implementation of pe_preinit for vmnet will follow in a separate
commit, to allow ordinary users to create VMs with vmnet
interfaces.

Signed-off-by: Steven McDonald <steven.mcdonald@usabilla.com>