Fix invalid-guest-state entry by updating state when CR0.PG cleared.
["cherry picked" from https://github.com/machyve/xhyve/commit/cc672d5363766e7c2bf10e02ca12efbeda74c487]

Recent Linux kernels (late 4.17) disable paging temporarily
when in long-mode. Handle this case by updating guest and VM
state to make sure they are in sync: EFER.LMA is cleared,
and the 'IA-32e mode guest' entry control is set to 0.

This is effectively the reverse operation to the existing case
when CR0.PG is set.

Signed-off-by: Rolf Neugebauer <rn@rneugeba.io>