trim: Delete unhelpful package.
This package is not helpful and no longer used by any known supported
importers. It's simpler and more readable to use strings.TrimSuffix and
strings.TrimPrefix, since package strings is in the standard library.

Any unknown importers of the trim package can be automatically rewritten
to use package strings instead with:

	gofmt -w -r 'trim.LastNewline(s) -> strings.TrimSuffix(s, "\n")' .
	gofmt -w -r 'trim.FirstSpace(s) -> strings.TrimPrefix(s, " ")' .
	goimports -w .