generated: remove backticks from specification text copy
The backticks were a part of Markdown in the original comment
to mark the regular expression as a code span. However, they're
redundant here, because that line is already indented, which is
another way of marking it as code.

There were no backticks in "Code generated" and "DO NOT EDIT." text
fragments, so don't add them here either. This part of the comment
gets displayed as a code block in godoc, and we want to stay more
true to the original specification text.