ioutil: delete unhelpful package
This package is not helpful and no longer used by any known supported
importers. It's simpler and more readable to use ioutil.WriteFile from
the standard library when a byte slice is available for writing.

If there's a real need to write directly from an io.Reader, then a
private helper (that can be customized as needed) is a better solution.

Also use go vet instead of go tool vet in .travis.yml. The latter has
been removed in favor of the former in Go 1.12.