On Majave and later, allocate with MAP_JIT
The VM requires writeable + executable memory which is blocked when
the "hardened runtime" is enabled at the codesigning stage.

If we pass the `MAP_JIT` flag then we can add the capability

  com.apple.security.cs.allow-jit [1]

to permit the allocation. The alternative is to globally enable
write+execute for all allocations in the process which seems worse [2]

[1] https://developer.apple.com/documentation/bundleresources/entitlements/com_apple_security_cs_allow-jit
[2] https://developer.apple.com/documentation/bundleresources/entitlements/com_apple_security_cs_allow-unsigned-executable-memory

Signed-off-by: David Scott <dave.scott@docker.com>