ocaml: handle unexpected error cases from the notifications pipe
The `Lwt_unix.read` is supposed to return 0 for EOF and >0 for number
of bytes read. Error results from the syscall are supposed to be handled
by the library. However there is evidence that an error `-1` is leaking
through, so add some logging and exit the process instead of spinning.

Signed-off-by: David Scott <dave@recoil.org>